Remove pending messages queue
Update the Server and Session classes to remove the pending message
queue from the Server class. Thus, it is not necessary to keep a
separated thread to process the queue. The thread used to receive the
message from the client will launch a task to process each message when
needed.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>