Using subscription request message
Updates the consumer program to use the SubscriptionRequestMessage class
to send the message subscribing to the given topic. This will make the
code cleaner.

Furthermore, the consumer program now reads the command line arguments
to know which are the server's address and port as well as the topic
which it should be using.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>